feat: styling of buttons
