+++ src/widget/auto_range_button.css
@@ -1,0 +1,27 @@
+/**
+ * @license
+ * Copyright 2024 Google Inc.
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+.neuroglancer-auto-range-button-container {
+  display: flex;
+  justify-content: flex-start;
+  margin-top: 2px;
+}
+
+.neuroglancer-auto-range-button {
+  font-size: 110%;
+  cursor: pointer;
+  margin-right: 4px;
+}
